requirements.txt
asgiref==3.9.1
charset-normalizer==3.4.3
colorama==0.4.6
Django==5.2.5
django-cors-headers==4.8.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
psycopg2-binary==2.9.10
pydot==4.0.1
PyJWT==2.10.1
pyparsing==3.2.3
python-decouple==3.8
python-dotenv==1.1.1
qrcode==8.2
reportlab==4.4.3
sqlparse==0.5.3
tzdata==2025.2
drf-yasg==1.21.7
